+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Fix calculation of how many KNN examples has the model in knn_forecasting
 * Weighted combination of the targets of nearest neighbors is implemented
+* A function that computes the number of training instances that would have 
+  a model has been added